[sysdig] Redirect to Agent Dashboard instead of Explore tab (#36)
* Sysdig Helm Chart output message

After deployment message updated. 

To check for a successful installation of the agent, we may point users to the new Dashboard Template for Agent Health & Status, instead of Explore.

Pls @nestorsalceda review this PR. Let me know if this is not the right way to do this! Thanks!

* fix: Increment versions and document it on CHANGELOG

This should fix Pablo's PR

Co-authored-by: pablopez <pablo.lopezzaldivar@sysdig.com>