[cloud-connector] allow role based access
The only current way to connect to the ingestors in AWS are through
a User and generate an access key and a secret.

However, kubernetes workloads can also assume a role at startup time
(for example, [https://github.com/jtblin/kube2iam](kube2iam)) without
the need to impersonate a user. IAM Users are limited and their secrets
are difficult to rotate. On the other hand, assuming an IAM role
generates temporary credentials at each assume and is safe to use.

In the current implementation, the secret will always include the AWS
credentials (even if empty) in the secret, and the deployment always
loads it in an environment variable at load time.

That causes a problem as in the AWS SDK, the env vars will take
precedence over the [role assumption](https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/credentials.html).
This PR introduces a new toggled called `aws.roleBasedAccess`, it
doesn't add the accessKeyId and secretAccessKey into the secret and it
doesn't mount it from the deployment, leaving the connector to assume
the IAM role.