Fix agent logPriority rendering
The initial implementation of the logPriority
flag mistakenly discarded anything other than
the console_priority and file_priority log
keys. Additionally, there was an indention
issue with whatever the subsequent line
following the 'log' items in the sysdig.settings
dictionary This change improves and simplifies
the handling of the log settings.

New tests have been created as well :)