revert sysdig chart agent version bump
I got too excited when I saw a PR for bumping the agent
version from 12.10.0 to 12.10.1 and enabled auto-merge on
the PR when I gave my review. That broke the release
pipeline as the PR was auto-merged before the pipeline
could merge it, and that caused the failure. This commit
will temporarily walk back the agent version in the sysdig
chart to 12.10.0 to enable the release job to run again.