Alphabetize tables in README files
To ease user experience when tuning the charts, this change alphabetizes the
parameters in the charts to make finding values much easier. Prior to this
change, finding a specific parameter in one of the larger READMEs, like
node-analyzer, was quite frustrating as the act of searching often require
the use of ctrl+f in the browser.