Fix issues with sysdig-deploy unit tests
It turns out that the GitHub unit test action was not running the tests for
the sysdig-deploy chart. As such, with the passing of development and bug fixes,
the unit tests that were written had gotten out of sync with the current state of
the charts. This change updates the tests so that they are passing again and
also adds the sysdig-deploy chart to the set of charts with unit tests run when
a PR is opened.