Create Delegated Agent installation mechanism
Add the ability to install the agent in a specialized delegated
manner where only a small portion of the overall nodes in the
cluster communicate with the Kubernetes API server. This is accomplished
by installing a modified Agent DaemonSet with an additional Agent
Deployment. The Agents in the Deployment are the ones which communicate
with the Kubernetes API server, but don't do much, if anything, else.