Add calls to "get_if_not_in_settings" back
A regression was discovered during testing where if the collector settings were
provided in the sysdig.settings block and not via the provided parameters
(collectorSettings.collectorHost, region, etc.) then we were rendering out the
collector information for the default region (us1) as well as the content
stored in the sysdig.settings block. Since the defined parameters are rendered
first, the unintended collector information for us1 appears before the values
specified in the settings block which results in the agent attempting to
connect to the wrong collectors. This change adds the safety net back that had
covered this workflow before that had been accidentally omitted.