Reduce mode misconfiguration from fail to NOTES output
The introduction of the monitor.enabled flag introduced a breaking change
for users explicitly putting the Agent into secure or secure_light mode by
informing the user that putting the Agent into one of those two modes while
running the Sysdig Monitor components is not advised. This change reduces the
intentional rendering fail into a suggestion in the NOTES.txt output.