Improve KubeVersion Comparisons
In various places across many charts we do compatibility checks against
exact Kubernetes versions. To do so, the charts had been casting the raw value
of `.Capabilities.KubeVersion.Minor` to an int like the following:

```
(lt (int .Capabilities.KubeVersion.Minor) 25)
```

The bug here is that certain Kubernetes flavors (EKS, microk8s, etc.) will
sometimes put a `+` at the end of the minor version. That `+` was causing the
cast to an int to silently fail and return a zero, which was causing failure
by attempting to deploy constructs that are not supported on incorrect cluster
versions.

This change guards against the `+` suffix as well as abstracts away the version
check logic into another helper function that can be more easily adjusted if
necessary.