fix(agent): Do not infer `max_parallel_cold_start` value
PR #1069 fixed a typo in the `max_parallel_cold_start` key. That fix resulted
in the inferred values being generated actually being used for the first time.
Looking through the commit history in the charts it looks like the typo had
been in place since the beginning, but now that the values are being consumed
by the agent they are resulting in undesirable behavior.

The fix here is to no longer infer or generate a number for the
`max_parallel_cold_start` key, but instead just rely on the Agent's default
unless the user specifies something else.