fix(sysdig-deploy): redo deployRuntimeScanner check to remove helm version constraint
The `deployRuntimeScanner` template function on `sysdig-deploy` was leveraging
some nuances of Helm
(https://github.com/helm/helm/issues/4535#issuecomment-416022809) that do not
reliably work for us with Helm versions <3.10. The issue comes from the import
of a template function from the node-analyzer chart when the user specifies
`nodeAnalyzer.enabled=false` and `clusterScanner.enabled=true` when using
`sysdig-deploy`. This change duplicates the `deployRuntimeScanner` template
logic from the `node-analyzer` chart into the `sysdig-deploy` chart to work
around the issue and remove the Helm version requirement.