feat(agent): lower default cpu requests for kmodule container
Quite often the value of 1000m for the kmodule CPU requests would cause
scheduling issues for the agent pod. This change lowers the requests to allow
scheduling to happen more often without needing to override these kmodule
settings. The limits will be kept the same however so that the container may
still burst to the same levels as before.