feat(agent,rapid-response): set metadata.namespace on all namespaced items
There were a handful of places in the charts related to `sysdig-deploy` where
the `metadata.namespace` field was not being explicitly set. This is works out
fine as Helm will set those fields itself during the `helm install` process,
but does not do so during a call to `helm template`. What has been discovered
is that for the workflow of generating manifest files via `helm template` and
then applying them later with `kubectl apply -f ...`, certain CI utilities will
patch the manifests not explicitly setting their namespace with a placeholder.
When the subsequent `kubectl apply -f ...` command is run that supplies the
desired namespace, the command fails because the namespace field is already set
on some constructs and is different from what is being requested. This change ensures that all namespaced items have the values of `metadata.namespace`
explicitly set to prevent the above issue.