fix(agent): don't duplicate `cointerface_enabled` field in deployment agent configmaps
The value `cointerface_enabled: true` is hardcoded into the deployment agent's configmap
by design, but if the value were overridden in `sysdig.settings` the field would end up
being duplicated when the deployment configmap blindly rendered out that block. This fix
addresses that issue by omitting that key when rendering the set of values from the
`sysdig.settings` block.