fix: prevent `custom_tags` from being set to `nil`
If the user did not set `global.sysdig.tags` when using `sysdig-deploy`
to install the host scanner, the `custom_tags` entry in the ConfigMap
was being set to `nil`. The node analyzer DaemonSet uses this entry to
set an environment variable on the host scanner pod(s), and a `nil` value
here was causing errors.