fix(agent): properly render security block content
There had been an issue where if the user set the
agent.secure.enabled flag to true in the charts, any
custom data in agent.sysdig.settings.security would
not be properly rendered. This change addresses the issue
by inspecting agent.sysdig.settings for any security block
content when the rest of the Sysdig Secure items are being
computed.