fix(agent): fix path of CA certificate to be used for proxy
We're currently mounting the custom CA certificate under
/opt/draios/etc/ca-certs as it is the location where the agent
should normally look for it (i.e. the root install /opt/draios).
However, the current agent implementation will always
treat ca_certificate as a relative path and always prepend
/opt/draios, whereas http_proxy.ca_certificate will honor
an absolute path if such a path is specified.
So if we specify /etc/ca-certs/... for http_proxy.ca_certificate,
any SSL connection to the proxy will fail because that path
will not exist. It will however work for the CA certificate
against the collector due to the reason explained above.

On the other hand, using a relative path etc/ca-certs will work
in both cases, so this is what we should be using instead.