Explicitly naming the prometheus metrics port in the agent daemonset
In order for the prometheus operator to scrape agent health metrics
from the agent, the port exposing these metrics must be named in the
daemonset's podspec. This commit will provide that port name when
agent.sysdig.settings.prometheus_exporter.enabled is true. It can
also account for non-default ports.

Testing: added a unit test; manually edited a daemonset in a live
cluster to ensure that the prometheus operator really does pick up the
metrics using a PodMonitor.