Adding checksum annotations to the clustershield deployment
Imagine the following scenario: the clustershield helm chart is
re-deployed and the contents of either its configmap, secrets, or
webhook change... and the clustershield deployment itself does not
change. The clustershield deployment won't restart because it does not
know that anything has changed - even though the things it depended on
change.

This problem is especially pronounced for clustershield because with
the default behavior it will auto-generate the certificates used for
communication between the webhook and the deployment. If the webhook
certificate changes and the deployment doesn't restart, the webhook
won't be able to communicate with the pods in the deployment.

The canonical way to address this is to put checksums of the
depended-upon templates into the deployment as annotations. Now when
resources the deployment depend upon change (like the auto-genereted
certificates) the clustershield pods will restart.

This is the same solution implemented in the nodeanalyzer daemonset.