fix(shield): reconfigure host shield configmap mounts
Mounting ConfigMaps while utilizing subPaths prevents the
mounts from being updated when the ConfigMap is updated.
See: https://github.com/kubernetes/kubernetes/issues/50345

Since all the files being mounted from the ConfigMap are
located in the same directory, the use of subPaths can
be removed and thus allow the configurations to be
properly picked up by the host shield when they are made.