feat(agent): [SMAGENT-8424] Allow for custom security context
Introduce a new top-level entry `customSecurityContext` in Values, so to
be able to override the securityContext automatically populated by the
chart. This is to be used sparingly, and particularly in circumstances
where a very harcoded securityContext should be used instead (like in
the case of BottleRocket).

Add unit tests for this feature and for the least privileged agent.

Adapt the documentation.